- Fixed starvation of low priority threads after penalty reschedule with a new reschedule function in heap module
- Fixes some compilation warnings
- Allow syscall definitions in any order by using C99 array initialization style

- Implementing missing syscall functions for dev subsystem
- Updating comments

+++ src/locks.c
@@ -17,7 +17,7 @@ uint32_t locks_syscall(uint32_t type, void * p)
         return (uint32_t)test_and_set((int *)p);
 
     default:
-        return NULL;
+        return (uint32_t)NULL;
     }
 }
